pyproject.toml
[tool.poetry]
name = "fastapi1"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
pydantic-extra-types = "^2.7.0"
pydantic = {extras = ["email"], version = "^2.7.1"}
black = "^24.4.2"
phonenumbers = "^8.13.36"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
